requirements.txt
ewmh ; sys_platform == "linux"
Pillow
PyAutoGUI
PyWinCtl
